cleaner ColorInputFieldWithPicker
